+++ preflight/src/main/java/org/apache/pdfbox/preflight/parser/PreflightParser.java
@@ -335,8 +335,10 @@ public class PreflightParser extends PDFParser
         while (true)
         {
             // just after the xref<EOL> there are an integer
-            long currObjID; // first obj id
-            long count; // the number of objects in the xref table
+            // first obj id
+            long currObjID;
+            // the number of objects in the xref table
+            int count; 
 
             long offset = pdfSource.getOffset();
             String line = readLine();
@@ -344,7 +346,7 @@ public class PreflightParser extends PDFParser
             Matcher matcher = pattern.matcher(line);
             if (matcher.matches())
             {
-                currObjID = Integer.parseInt(matcher.group(1));
+                currObjID = Long.parseLong(matcher.group(1));
                 count = Integer.parseInt(matcher.group(2));
             }
             else
@@ -353,8 +355,10 @@ public class PreflightParser extends PDFParser
                         "Cross reference subsection header is invalid"));
                 // reset pdfSource cursor to read xref information
                 pdfSource.seek(offset);
-                currObjID = readObjectNumber(); // first obj id
-                count = readLong(); // the number of objects in the xref table
+                // first obj id
+                currObjID = readObjectNumber();
+                // the number of objects in the xref table
+                count = readInt();
             }
 
             skipSpaces();
@@ -379,9 +383,7 @@ public class PreflightParser extends PDFParser
                             "invalid xref line: " + currentLine));
                     break;
                 }
-                /*
-                 * This supports the corrupt table as reported in PDFBOX-474 (XXXX XXX XX n)
-                 */
+                // This supports the corrupt table as reported in PDFBOX-474 (XXXX XXX XX n)
                 if (splitString[splitString.length - 1].equals("n"))
                 {
                     try
@@ -613,8 +615,7 @@ public class PreflightParser extends PDFParser
     }
 
     @Override
-    protected COSBase parseObjectDynamically(long objNr, int objGenNr,
-            boolean requireExistingNotCompressedObj)
+    protected COSBase parseObjectDynamically(long objNr, int objGenNr, boolean requireExistingNotCompressedObj)
             throws IOException
     {
         // ---- create object key and get object (container) from pool
@@ -662,7 +663,7 @@ public class PreflightParser extends PDFParser
                 Matcher matcher = pattern.matcher(line);
                 if (matcher.matches())
                 {
-                    readObjNr = Integer.parseInt(matcher.group(1));
+                    readObjNr = Long.parseLong(matcher.group(1));
                     readObjGen = Integer.parseInt(matcher.group(2));
                 }
                 else
